Add test_mutable.Problems.test_multiply_placed_shares
The UCWEs in the incident reports associated with #1628 all seem to be
associated with shares that the servermap knows about, but which aren't
accounted for during the publish process for whatever reason.
Specifically, it looks like the publisher is only capable of keeping
track of a single storage server for a given share. This makes the
repair process worse than it was pre-MDMF at updating all of the shares
of a particular file to the newest version, and can also cause spurious
UCWEs. This test simulates such a layout and fails if an UCWE is thrown.
We need to write another test to ensure that all copies of a share are
updated to the latest version (or alter this test to do that), so that
the test suite doesn't pass unless both regressions are fixed.