Change self.writers to map a shnum to a list of writers.
Before this commit, the publisher only kept track of a single writer for
each share. This is insufficient to handle updates in which a single
share may live on multiple servers. In the best case, an update will
only update one of the existing shares instead of all of them. In some
cases, the update will encounter the existing shares when publishing
some other share, interpret it as a sign of an uncoordinated update, and
fail. Keeping track of all of the writers helps ensure that all existing
shares are updated, and helps avoid spurious uncoordinated write errors.