Add NoNetworkGrid.copy_mutable_share_to_server
With immutable shares, we can just deal with the files on disk. We can't
do that with mutable shares, since the write enabler is bound to the
nodeid of the storage server storing the share. This routine uses
MutableShareFiles to copy the actual mutable share and leases in a way
that yields a correct write enabler. A naive file copy would yield a
share that's not useful to a client, since any attempt to access it
would result in a write enabler mismatch and an error.