Rename get_checkstring to get_next_checkstring.
get_next_checkstring reflects the relationship of the checkstring
returned by the method to what's actually on the server. In particular,
it draws a distinction between what's there now and what's going to be
there when the write takes place that isn't drawn now, leading to
confusion when reading the code.