Make UCWE exemptions more conservative.
The code removed by this commit assumed that any surprise shares
associated with concurrent writes to the same server weren't surprise
shares. This is a regression relative to the old publisher, which
checked the read vectors against the publish operation's checkstring
before ignoring the surprise shares. We can't do this with the new
publisher, since there's not a guaranteed write order. Instead, we allow
two checkstrings. If the readv is the same as the publish operation's
checkstring, then the share is there due either to a convergent
uncoordinated write or an earlier writer finishing before the one being
responded to. If the readv is the same as the checkstring we saw during
the servermap update, then the share hasn't been updated yet. In these
cases, the surprise shares aren't evidence of an uncoordinated write and
can be ignored.