Add a test for the re-opened #1212.
This test builds a grid of four servers, then uploads a file with
encoding parameters k=2, n=4, happy=4, then deletes a share, then
repairs the file. According to #1212, this should result in an
UploadUnhappinessError because it relies on the value of happy defined
in BaseUploadable, which is 7. The test ensures that this isn't the
case, and that the file is repaired successfully. The test also passes
without any modifications, which suggests that I'm missing something.