Add _do_inquiry method; rework readonly check to use _do_inquiry.
do_inquiry contacts a storage server (through its tracker instance)
asking about existing shares, recording any that it finds to our
internal state. In the current codebase, this will be used for the
readonly check, but we'll be able to use it later to handle the inquiry
tasks given to us by the peer selector.