Make uncoordinated write check in mutable publisher more conservative.
The old check (removed by this commit) assumed that any surprise shares
found on a server that could be explained by this write were due to
this write, presumably to avoid being surprised by our own write
operation. This is not particularly conservative, and may not detect
uncoordinated writes. For example, if publisher A and publisher B are
performing simultaneous divergent uncoordinated updates to some file C,
and both opt to follow the existing share placement (as publishers
prefer to do), the existing check will A to ignore B's new shares, and
vice-versa, eliminating their ability to detect the uncoordinated write.

The new check explicitly deals with the two cases in which we want to
ignore surprise shares. First, we allow shares that have a checkstring
associated with the mutable file version being written by the current
publish operation. Second, we allow shares that have a checkstring
associated with the version that we're overwriting or updating. All
other shares are interpreted as evidence of an uncoordinated write. This
resolves the issue described above.