Rename checkstring methods; add method to get existing checkstring
get_existing_checkstring and set_existing_checkstring deal with the
checkstring that the writer expects to exist on the remote server.
get_next_checkstring returns the checkstring that will exist on the
remote server after the next write (i.e., the checkstring associated
with the version of the mutable file that is about to be published).

The new names make the purpose of the functions clearer at a glance.